requirements.txt
customtkinter>=5.2
ultralytics>=8.0
face-recognition>=1.3
opencv-python-headless>=4.8
Pillow>=10.0
pandas>=2.0
numpy>=1.24
